+++ src/game_room/game_room_client.py
@@ -2,6 +2,8 @@ from sanic.server.websockets.impl import WebsocketImplProtocol
 from sanic.request import Request
 from typing import Final
 from sanic import Sanic
+from sanic.response import text
+
 
 app: Final = Sanic("GameCodin", log_config={"version": 1})
 
@@ -13,4 +15,13 @@ app.config.WEBSOCKET_PING_TIMEOUT = None  # type: ignore
 
 @app.websocket("/", name='ws')
 async def ws_handler(request: Request, ws: WebsocketImplProtocol):
-    pass
+    await ws.send("pog")
+
+
+@app.get('/users')
+async def handler(request):
+
+    return text("Ok")
+
+
+app.run(host="0.0.0.0", port=8080, workers=1, debug=False, access_log=False)
